requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
openpyxl
python-multipart
python-jose[cryptography]
bcrypt
python-dotenv
Pillow
